changed reading tmp/hum
created one function for tmp/hum reading throughout app.
changed how often is readed the real temp and hum. if there is another read in the past 5 mins, this is displayed to user or used for DB data, else a new read is requested to the reader process.
room manager still in alpha.